[IMG-116] Add test for 8 bit mono image with bitshift.
Also fix bug that this test showed up.

The test data is 7 bit ABPP in 8 bit NBPP.

+++ render/src/main/java/org/codice/imaging/nitf/render/datareader/Bitshift8IOReaderFunction.java
@@ -41,7 +41,7 @@ class Bitshift8IOReaderFunction implements IOReaderFunction {
 
     @Override
     public Object apply(final Object imageInputStream) throws IOException {
-        return (((ImageInputStream) imageInputStream).readBits(this.bitsToRead) << this.bitShift);
+        return (int) (((ImageInputStream) imageInputStream).readBits(this.bitsToRead) << this.bitShift);
     }
 
 }
